fix angularError and lint

+++ PAMI/lib/rolling_basis/include/point.h
@@ -34,10 +34,10 @@ public:
 
     static float distance(Point p1, Point p2)
     {
-        return sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2));
+        return sqrtf(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2));
     }
     static float angle(Point p1, Point p2)
     {
-        return atan2(p2.y - p1.y, p2.x - p1.x);
+        return atan2f(p2.y - p1.y, p2.x - p1.x);
     }
 };
